Fix some issues
