+++ docs/source/index.rst
@@ -3,20 +3,15 @@ Homework: Session 17-19 Documentation
 
 Documentation for sessions 17 to 19 Homework assignment.
 
-----------
 Session 17
 ----------
 
-========================
-.. doxygennamespace:: homework17
-   :project: Homework
-   :members:
 
-----------
+.. doxygennamespace:: homework17
+
+
 Session 19
 ----------
 
-========================
+
 .. doxygennamespace:: homework19
-   :project: Homework
-   :members:
